modules/location-storage-writer/location-storage-writer/requirements.txt
grpc-interceptor ~= 0.11.0
grpcio-tools ~= 1.30 
confluent-kafka==1.9.0